+++ cli/src/index.test.mjs
@@ -1,18 +1,62 @@
-import * as nc from 'newcar'
+import path from 'node:path'
+import { pathToFileURL } from 'node:url'
+import { CarEngine, Color, ImageWidget, Scene, Text, Widget, useFont, useImage } from 'newcar'
 
-const engine = await new nc.CarEngine().init(
-  '../cli/node_modules/canvaskit-wasm/bin/canvaskit.wasm',
-)
+const logoLoaded = useImage('./assets/newcar.svg')
+await useFont(path.resolve('./fonts/bahnschrift.ttf'))
+
+const engine = await new CarEngine().init('../node_modules/canvaskit-wasm/bin/canvaskit.wasm')
 const app = engine.createLocalApp(1600, 900)
-const root = new nc.Circle(100, {
-  x: 800,
-  y: 100,
-}).animate(nc.create, 0, 60).animate(nc.move, 60, 60, {
-  from: [800, 100],
-  to: [800, 800],
-  by: nc.easeBounce,
+const logo = new ImageWidget(logoLoaded, {
+  x: 668,
+  y: 386
 })
-const scene = new nc.Scene(root)
+const text = new Text([{
+  text: "Hello! ",
+  style: {
+    fontSize: 50,
+    fontFamilies: ['bahnschrift']
+  }
+}, {
+  text: "N",
+  style: {
+    fontSize: 40,
+    color: Color.parse('red')
+  }
+}, {
+  text: "e",
+  style: {
+    fontSize: 60,
+    color: Color.parse('orange')
+  }
+},{
+  text: "w",
+  style: {
+    fontSize: 30,
+    color: Color.parse('yellow')
+  }
+},{
+  text: "c",
+  style: {
+    fontSize: 50,
+    color: Color.parse('green')
+  }
+},{
+  text: "a",
+  style: {
+    fontSize: 30,
+    color: Color.parse('blue')
+  }
+},{
+  text: "r",
+  style: {
+    fontSize: 40,
+    color: Color.parse('purple')
+  }
+}], {
+})
+const root = new Widget().add(logo, text)
+const scene = new Scene(root)
 app.checkout(scene)
 
 export default app