+++ mods/mod-geometry/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @newcar/mod-geometry
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @newcar/basic@0.14.2
+  - @newcar/core@0.14.2
+  - @newcar/utils@0.14.2
+
 ## 0.1.2
 
 ### Patch Changes