+++ packages/basic/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @newcar/basic
 
+## 0.14.2
+
+### Patch Changes
+
+- fix: failed to load sources under local mode.
+- Updated dependencies
+  - @newcar/core@0.14.2
+  - @newcar/utils@0.14.2
+
 ## 0.14.1
 
 ### Patch Changes